+++ packages/aim_postgres/lib/aim_postgres.dart
@@ -36,4 +36,4 @@ library;
 
 export 'src/pg_connection.dart';
 export 'src/pg_database.dart';
-export 'src/pool/pool.dart';
+export 'src/pool/pool.dart' show PoolOptions, PoolStats, PoolTimeoutException;